Correct header file order, move the model validation code to ila folder
