+++ include/ilang/ila/validate_model.h
@@ -8,9 +8,9 @@
 namespace ilang {
 enum DEFAULT_UPDATE_METHOD { OLD_VALUE = 0, NONDET_VALUE = 1 };
 
-bool deterministic_check(const InstrLvlAbsPtr& model_ptr_);
-bool complete_check(const InstrLvlAbsPtr& model_ptr_);
-void complete_model(const InstrLvlAbsPtr& model_ptr_,
+bool CheckDeterminism(const InstrLvlAbsPtr& model_ptr_);
+bool CheckCompleteness(const InstrLvlAbsPtr& model_ptr_);
+void CompleteModel(const InstrLvlAbsPtr& model_ptr_,
                     DEFAULT_UPDATE_METHOD dum);
 } // namespace ilang
